Initial Commit for 'data-structures' repo
First commit of the 'data-structures' repository for my own practice
attempts at more complex C++ features/concepts.

This commit adds 'include/utilities.hpp', which contains a handful of
using/type aliases for simplicity as well as a couple structures to
represent typeclasses that implement either '>' (Ord) or '==' (Eq). The
utilities header is complete for the time being. The second file in this
commit is 'include/Node.hpp', which is mostly incomplete but will
eventually contain the class template for the 'Node' type.